+++ CHANGELOG.md
@@ -3,6 +3,14 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 This CHANGELOG follows the format listed at [Keep A Changelog](http://keepachangelog.com/)
 
+## 0.0.3 - 2016-11-24
+### Added
+- Virtual Network Gateways:
+  - Added functionality to check Failover connections
+  - Made it possible to specify a custom Scheme for Metrics
+- Service Bus:
+  - Added a Metric to expose the the Message Count for a Service Bus Topic
+
 ## 0.0.2 - 2016-11-23
 ### Added
 - Virtual Network Gateways: